tests: switch errors to warnings in test
