+++ crates/pixi_manifest/src/toml/manifest.rs
@@ -550,10 +550,12 @@ impl<'de> toml_span::Deserialize<'de> for TomlManifest {
 #[cfg(test)]
 mod test {
     use insta::assert_snapshot;
-    use itertools::Itertools;
 
     use super::*;
-    use crate::{toml::FromTomlStr, utils::test_utils::format_parse_error};
+    use crate::{
+        toml::FromTomlStr,
+        utils::test_utils::{expect_parse_warnings, format_parse_error},
+    };
 
     /// A helper function that generates a snapshot of the error message when
     /// parsing a manifest TOML. The error is returned.
@@ -566,21 +568,6 @@ mod test {
             .expect_err("parsing should fail");
 
         format_parse_error(pixi_toml, parse_error)
-    }
-
-    /// A helper function that generates a snapshot of the warnings message when
-    /// parsing a manifest TOML. The error is returned.
-    #[must_use]
-    pub(crate) fn expect_parse_warnings(pixi_toml: &str) -> String {
-        match <TomlManifest as FromTomlStr>::from_toml_str(pixi_toml).and_then(|manifest| {
-            manifest.into_workspace_manifest(ExternalWorkspaceProperties::default(), None)
-        }) {
-            Ok((_, _, warnings)) => warnings
-                .into_iter()
-                .map(|warning| format_parse_error(pixi_toml, warning))
-                .join("\n\n"),
-            Err(err) => format_parse_error(pixi_toml, err),
-        }
     }
 
     #[test]
@@ -749,9 +736,10 @@ mod test {
 
     #[test]
     fn test_target_workspace_dependencies() {
-        assert_snapshot!(expect_parse_failure(
+        assert_snapshot!(expect_parse_warnings(
             r#"
         [workspace]
+        name = "test"
         channels = []
         platforms = ['osx-64']
         preview = ["pixi-build"]
@@ -768,9 +756,10 @@ mod test {
 
     #[test]
     fn test_mismatching_target_selector() {
-        assert_snapshot!(expect_parse_failure(
+        assert_snapshot!(expect_parse_warnings(
             r#"
         [workspace]
+        name = "test"
         channels = []
         platforms = ['win-64']
 
@@ -781,9 +770,10 @@ mod test {
 
     #[test]
     fn test_mismatching_multi_target_selector() {
-        assert_snapshot!(expect_parse_failure(
+        assert_snapshot!(expect_parse_warnings(
             r#"
         [workspace]
+        name = "test"
         channels = []
         platforms = ['win-64']
